+++ app/src/main/java/black/bracken/neji/ui/regionlist/RegionListViewModel.kt
@@ -1,4 +1,4 @@
-package black.bracken.neji.ui.top
+package black.bracken.neji.ui.regionlist
 
 import androidx.hilt.Assisted
 import androidx.hilt.lifecycle.ViewModelInject
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import black.bracken.neji.repository.FirebaseRepository
 
-class TopViewModel @ViewModelInject constructor(
+class RegionListViewModel @ViewModelInject constructor(
     @Assisted private val savedState: SavedStateHandle,
     private val firebaseRepository: FirebaseRepository
 ) : ViewModel() {
