+++ app/src/main/res/layout/region_list_card.xml
