+++ app/src/main/res/values/strings.xml
@@ -6,9 +6,8 @@
     <string name="setup_hint_email">Email</string>
     <string name="setup_hint_password">Password</string>
     <string name="setup_text_login">Login</string>
-    <string name="top_success_sign_in">ログインに成功しました</string>
-    <string name="top_card_region_name">リージョン: %s</string>
-    <string name="top_card_region_box_amount">ボックス数: %d個</string>
+    <string name="region_list_card_region_name">リージョン: %s</string>
+    <string name="region_list_card_box_amount">ボックス数: %d個</string>
     <string name="add_item_name">アイテム名</string>
     <string name="add_item_amount">個数</string>
     <string name="add_item_type">アイテムの種類</string>
@@ -18,8 +17,9 @@
     <string name="add_item_adding">追加</string>
     <string name="error_must_not_be_blank">値を入力してください</string>
     <string name="error_must_be_integer_and_at_least_zero">0以上の整数にしてください</string>
+    <string name="snackbar_success_sign_in">ログインに成功しました</string>
     <string name="snackbar_invalid_value">エラー: 不正な値が存在します</string>
     <string name="snackbar_must_not_be_blank">エラー: 空欄が存在します</string>
-    <string name="nav_name_top">場所一覧</string>
+    <string name="nav_name_region_list">場所一覧</string>
     <string name="nav_name_add_item">アイテムの追加</string>
 </resources>
